+++ src/core/converter/ImportFileAeroflyCustomMissionsTmcConverter.ts
@@ -2,7 +2,7 @@ import { AeroflyCustomMissionsParser } from "./AeroflyCustomMissionsParser.js";
 import { ImportFileConverter } from "./ImportFileConverter.js";
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
 
-export class ImportFileAeroflyCustomMissionsTmcConverter implements ImportFileConverter {
+export class ImportFileAeroflyCustomMissionsTmcConverter extends ImportFileConverter {
     static readonly fileExtension = "tmc";
 
     convert(content: string, flightplan: AeroflyFlight): void {