+++ types/core/converter/ImportFileAeroflyCustomMissionsTmcConverter.d.ts
@@ -1,6 +1,6 @@
 import { ImportFileConverter } from "./ImportFileConverter.js";
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
-export declare class ImportFileAeroflyCustomMissionsTmcConverter implements ImportFileConverter {
+export declare class ImportFileAeroflyCustomMissionsTmcConverter extends ImportFileConverter {
     static readonly fileExtension = "tmc";
     convert(content: string, flightplan: AeroflyFlight): void;
 }