+++ types/core/converter/ImportFileAeroflyMcfConverter.d.ts
@@ -1,6 +1,6 @@
 import { ImportFileConverter } from "./ImportFileConverter.js";
 import { AeroflyFlight } from "@fboes/aerofly-custom-missions";
-export declare class ImportFileAeroflyMcfConverter implements ImportFileConverter {
+export declare class ImportFileAeroflyMcfConverter extends ImportFileConverter {
     static readonly fileExtension = "mcf";
     convert(content: string, flightplan: AeroflyFlight): void;
 }